requirements.txt
# Core
aiohttp>=3.9.0
asyncio>=3.4.3
python-dotenv>=1.0.0
pyyaml>=6.0
schedule>=1.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tenacity>=8.2.0

# GitHub
PyGithub>=2.3.0
gitpython>=3.1.40

# Logging
loguru>=0.7.2